feat(log): LogPage `Suspense` 적용 (#220)
